+++ Runtime/SystemMetricsMali.cs
@@ -12,9 +12,13 @@ namespace Unity.Profiling.SystemMetrics
         public bool Active { get { return SystemMetricsLibHWCP.IsActive(); } }
 
         public ProfilerRecorderHandle GpuActive { get; private set; }
+        public ProfilerRecorderHandle GpuShaderCoreActive { get; private set; }
         public ProfilerRecorderHandle GpuNonFragmentActive { get; private set; }
         public ProfilerRecorderHandle GpuFragmentActive { get; private set; }
         public ProfilerRecorderHandle GpuTilerActive { get; private set; }
+        public ProfilerRecorderHandle GpuShaderCoreUtilization { get; private set; }
+        public ProfilerRecorderHandle GpuNonFragmentUtilization { get; private set; }
+        public ProfilerRecorderHandle GpuFragmentUtilization { get; private set; }
 
         public ProfilerRecorderHandle GpuVertexComputeJobs { get; private set; }
         public ProfilerRecorderHandle GpuTiles { get; private set; }
@@ -30,19 +34,37 @@ namespace Unity.Profiling.SystemMetrics
         public ProfilerRecorderHandle GpuInstructions { get; private set; }
         public ProfilerRecorderHandle GpuDivergedInstructions { get; private set; }
 
-        public ProfilerRecorderHandle GpuShaderCycles { get; private set; }
-        public ProfilerRecorderHandle GpuShaderArithmeticCycles { get; private set; }
-        public ProfilerRecorderHandle GpuShaderLoadStoreCycles { get; private set; }
-        public ProfilerRecorderHandle GpuShaderTextureCycles { get; private set; }
+        public ProfilerRecorderHandle GpuShaderArithmeticActive { get; private set; }
+        public ProfilerRecorderHandle GpuShaderLoadStoreActive { get; private set; }
+        public ProfilerRecorderHandle GpuShaderTextureUnitActive { get; private set; }
+        public ProfilerRecorderHandle GpuShaderFragmentActive { get; private set; }
+        public ProfilerRecorderHandle GpuShaderComputeActive { get; private set; }
+        public ProfilerRecorderHandle GpuShaderTripipeActive { get; private set; }
+        public ProfilerRecorderHandle GpuShaderFragmentUtilization { get; private set; }
+        public ProfilerRecorderHandle GpuShaderComputeUtilization { get; private set; }
+        public ProfilerRecorderHandle GpuShaderTripipeUtilization { get; private set; }
+        public ProfilerRecorderHandle GpuShaderArithmeticUtilization { get; private set; }
+        public ProfilerRecorderHandle GpuShaderLoadStoreUtilization { get; private set; }
+        public ProfilerRecorderHandle GpuShaderTextureUtilization { get; private set; }
 
         public ProfilerRecorderHandle GpuCacheReadLookups { get; private set; }
         public ProfilerRecorderHandle GpuCacheWriteLookups { get; private set; }
-        public ProfilerRecorderHandle GpuExternalMemoryReadAccesses { get; private set; }
-        public ProfilerRecorderHandle GpuExternalMemoryWriteAccesses { get; private set; }
-        public ProfilerRecorderHandle GpuExternalMemoryReadStalls { get; private set; }
-        public ProfilerRecorderHandle GpuExternalMemoryWriteStalls { get; private set; }
-        public ProfilerRecorderHandle GpuExternalMemoryReadBytes { get; private set; }
-        public ProfilerRecorderHandle GpuExternalMemoryWriteBytes { get; private set; }
+        public ProfilerRecorderHandle GpuMemoryReadAccesses { get; private set; }
+        public ProfilerRecorderHandle GpuMemoryWriteAccesses { get; private set; }
+        public ProfilerRecorderHandle GpuMemoryReadStalls { get; private set; }
+        public ProfilerRecorderHandle GpuMemoryWriteStalls { get; private set; }
+        public ProfilerRecorderHandle GpuMemoryReadBytes { get; private set; }
+        public ProfilerRecorderHandle GpuMemoryWriteBytes { get; private set; }
+        public ProfilerRecorderHandle GpuReadStallRate { get; private set; }
+        public ProfilerRecorderHandle GpuWriteStallRate { get; private set; }
+
+        public ProfilerRecorderHandle GpuInputPrimitives { get; private set; }
+        public ProfilerRecorderHandle GpuCulledPrimitives { get; private set; }
+        public ProfilerRecorderHandle GpuClippedPrimitives { get; private set; }
+        public ProfilerRecorderHandle GpuVisiblePrimitives { get; private set; }
+        public ProfilerRecorderHandle GpuCulledPrimitivesPercentage { get; private set; }
+        public ProfilerRecorderHandle GpuClippedPrimitivesPercentage { get; private set; }
+        public ProfilerRecorderHandle GpuVisiblePrimitivesPercentage { get; private set; }
 
         static SystemMetricsMali()
         {
@@ -62,9 +84,13 @@ namespace Unity.Profiling.SystemMetrics
 
             // Set properties only if initialization is successeful
             GpuActive = SystemMetricsLibHWCP.GetProfilerRecorderHandle(nameof(GpuActive));
+            GpuShaderCoreActive = SystemMetricsLibHWCP.GetProfilerRecorderHandle(nameof(GpuShaderCoreActive));
             GpuNonFragmentActive = SystemMetricsLibHWCP.GetProfilerRecorderHandle(nameof(GpuNonFragmentActive));
             GpuFragmentActive = SystemMetricsLibHWCP.GetProfilerRecorderHandle(nameof(GpuFragmentActive));
             GpuTilerActive = SystemMetricsLibHWCP.GetProfilerRecorderHandle(nameof(GpuTilerActive));
+            GpuShaderCoreUtilization = SystemMetricsLibHWCP.GetProfilerRecorderHandle(nameof(GpuShaderCoreUtilization));
+            GpuNonFragmentUtilization = SystemMetricsLibHWCP.GetProfilerRecorderHandle(nameof(GpuNonFragmentUtilization));
+            GpuFragmentUtilization = SystemMetricsLibHWCP.GetProfilerRecorderHandle(nameof(GpuFragmentUtilization));
 
             GpuVertexComputeJobs = SystemMetricsLibHWCP.GetProfilerRecorderHandle(nameof(GpuVertexComputeJobs));
             GpuTiles = SystemMetricsLibHWCP.GetProfilerRecorderHandle(nameof(GpuTiles));
@@ -80,19 +106,37 @@ namespace Unity.Profiling.SystemMetrics
             GpuInstructions = SystemMetricsLibHWCP.GetProfilerRecorderHandle(nameof(GpuInstructions));
             GpuDivergedInstructions = SystemMetricsLibHWCP.GetProfilerRecorderHandle(nameof(GpuDivergedInstructions));
 
-            GpuShaderCycles = SystemMetricsLibHWCP.GetProfilerRecorderHandle(nameof(GpuShaderCycles));
-            GpuShaderArithmeticCycles = SystemMetricsLibHWCP.GetProfilerRecorderHandle(nameof(GpuShaderArithmeticCycles));
-            GpuShaderLoadStoreCycles = SystemMetricsLibHWCP.GetProfilerRecorderHandle(nameof(GpuShaderLoadStoreCycles));
-            GpuShaderTextureCycles = SystemMetricsLibHWCP.GetProfilerRecorderHandle(nameof(GpuShaderTextureCycles));
+            GpuShaderArithmeticActive = SystemMetricsLibHWCP.GetProfilerRecorderHandle(nameof(GpuShaderArithmeticActive));
+            GpuShaderLoadStoreActive = SystemMetricsLibHWCP.GetProfilerRecorderHandle(nameof(GpuShaderLoadStoreActive));
+            GpuShaderTextureUnitActive = SystemMetricsLibHWCP.GetProfilerRecorderHandle(nameof(GpuShaderTextureUnitActive));
+            GpuShaderFragmentActive = SystemMetricsLibHWCP.GetProfilerRecorderHandle(nameof(GpuShaderFragmentActive));
+            GpuShaderComputeActive = SystemMetricsLibHWCP.GetProfilerRecorderHandle(nameof(GpuShaderComputeActive));
+            GpuShaderTripipeActive = SystemMetricsLibHWCP.GetProfilerRecorderHandle(nameof(GpuShaderTripipeActive));
+            GpuShaderFragmentUtilization = SystemMetricsLibHWCP.GetProfilerRecorderHandle(nameof(GpuShaderFragmentUtilization));
+            GpuShaderComputeUtilization = SystemMetricsLibHWCP.GetProfilerRecorderHandle(nameof(GpuShaderComputeUtilization));
+            GpuShaderTripipeUtilization = SystemMetricsLibHWCP.GetProfilerRecorderHandle(nameof(GpuShaderTripipeUtilization));
+            GpuShaderArithmeticUtilization = SystemMetricsLibHWCP.GetProfilerRecorderHandle(nameof(GpuShaderArithmeticUtilization));
+            GpuShaderLoadStoreUtilization = SystemMetricsLibHWCP.GetProfilerRecorderHandle(nameof(GpuShaderLoadStoreUtilization));
+            GpuShaderTextureUtilization = SystemMetricsLibHWCP.GetProfilerRecorderHandle(nameof(GpuShaderTextureUtilization));
 
             GpuCacheReadLookups = SystemMetricsLibHWCP.GetProfilerRecorderHandle(nameof(GpuCacheReadLookups));
             GpuCacheWriteLookups = SystemMetricsLibHWCP.GetProfilerRecorderHandle(nameof(GpuCacheWriteLookups));
-            GpuExternalMemoryReadAccesses = SystemMetricsLibHWCP.GetProfilerRecorderHandle(nameof(GpuExternalMemoryReadAccesses));
-            GpuExternalMemoryWriteAccesses = SystemMetricsLibHWCP.GetProfilerRecorderHandle(nameof(GpuExternalMemoryWriteAccesses));
-            GpuExternalMemoryReadStalls = SystemMetricsLibHWCP.GetProfilerRecorderHandle(nameof(GpuExternalMemoryReadStalls));
-            GpuExternalMemoryWriteStalls = SystemMetricsLibHWCP.GetProfilerRecorderHandle(nameof(GpuExternalMemoryWriteStalls));
-            GpuExternalMemoryReadBytes = SystemMetricsLibHWCP.GetProfilerRecorderHandle(nameof(GpuExternalMemoryReadBytes));
-            GpuExternalMemoryWriteBytes = SystemMetricsLibHWCP.GetProfilerRecorderHandle(nameof(GpuExternalMemoryWriteBytes));
+            GpuMemoryReadAccesses = SystemMetricsLibHWCP.GetProfilerRecorderHandle(nameof(GpuMemoryReadAccesses));
+            GpuMemoryWriteAccesses = SystemMetricsLibHWCP.GetProfilerRecorderHandle(nameof(GpuMemoryWriteAccesses));
+            GpuMemoryReadStalls = SystemMetricsLibHWCP.GetProfilerRecorderHandle(nameof(GpuMemoryReadStalls));
+            GpuMemoryWriteStalls = SystemMetricsLibHWCP.GetProfilerRecorderHandle(nameof(GpuMemoryWriteStalls));
+            GpuMemoryReadBytes = SystemMetricsLibHWCP.GetProfilerRecorderHandle(nameof(GpuMemoryReadBytes));
+            GpuMemoryWriteBytes = SystemMetricsLibHWCP.GetProfilerRecorderHandle(nameof(GpuMemoryWriteBytes));
+            GpuReadStallRate = SystemMetricsLibHWCP.GetProfilerRecorderHandle(nameof(GpuReadStallRate));
+            GpuWriteStallRate = SystemMetricsLibHWCP.GetProfilerRecorderHandle(nameof(GpuWriteStallRate));
+
+            GpuInputPrimitives = SystemMetricsLibHWCP.GetProfilerRecorderHandle(nameof(GpuInputPrimitives));
+            GpuCulledPrimitives = SystemMetricsLibHWCP.GetProfilerRecorderHandle(nameof(GpuCulledPrimitives));
+            GpuClippedPrimitives = SystemMetricsLibHWCP.GetProfilerRecorderHandle(nameof(GpuClippedPrimitives));
+            GpuVisiblePrimitives = SystemMetricsLibHWCP.GetProfilerRecorderHandle(nameof(GpuVisiblePrimitives));
+            GpuCulledPrimitivesPercentage = SystemMetricsLibHWCP.GetProfilerRecorderHandle(nameof(GpuCulledPrimitivesPercentage));
+            GpuClippedPrimitivesPercentage = SystemMetricsLibHWCP.GetProfilerRecorderHandle(nameof(GpuClippedPrimitivesPercentage));
+            GpuVisiblePrimitivesPercentage = SystemMetricsLibHWCP.GetProfilerRecorderHandle(nameof(GpuVisiblePrimitivesPercentage));
         }
 
         static bool InstallUpdateCallback()