+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 2021-06-16
+
+### Support for more hardware counters and UX improvements
+- Updated HWCP library to the latest version.
+- Added custom Profiler Module "Mali System Metrics" with improved detailed view,
+- Added more hardware and derivated counters. See "Available metrics" for more details.
+- Improved metrics description documentation
+
 ## [0.1.0] - 2021-03-01
 
 ### Initial release